parseProgram moved to ParseProgram in cxparsering
